servidor_recepcion_datos/requirements.txt
Flask~=2.1.2
pymongo==3.12.1
kafka-python==2.0.2
jsonschema==3.2.0
Flask-session
pandas~=1.3.5
ruamel-yaml
markupsafe==2.0.1
PyYAML~=6.0
escape
flask_login
Werkzeug~=2.1.2
Jinja2~=3.1.2
click~=8.1.3
itsdangerous~=2.1.2
greenlet~=1.1.2
SQLAlchemy~=1.4.36
Flask-SQLAlchemy
psycopg2-binary
python-dotenv